feat: 총 수익률 계산
* Test 케이스 클래스 구분
